Rename DeliveratorApiClient to ApiClient
