+++ src/protagonist/API.Client/DlcsClient.cs
@@ -217,11 +217,6 @@ public class DlcsClient : IDlcsClient
         return patched;
     }
 
-    public Task<bool> ReingestAsset(int requestSpaceId, string requestImageId)
-    {
-        throw new NotImplementedException();
-    }
-
     private HttpContent ApiBody(JsonLdBase apiObject)
     {
         var jsonString = JsonConvert.SerializeObject(apiObject, jsonSerializerSettings);
