+++ src/protagonist/API/Features/Image/ImageController.cs
@@ -169,7 +169,6 @@ public class ImageController : HydraController
     public Task<IActionResult> ReingestAsset([FromRoute] int customerId, [FromRoute] int spaceId,
         [FromRoute] string imageId, CancellationToken cancellationToken)
     {
-        // TODO - return the { "success": true } response for backward compat
         var reingestRequest = new ReingestAsset(customerId, spaceId, imageId);
         return HandleUpsert(reingestRequest, 
             asset => asset.ToHydra(GetUrlRoots()),
